+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.0.1
 
 ### Updated
+- Updated `/projects/create-project` and added new shared RadioGroup and CheckboxGroup components [#332]
 - Updated `/projects/page.tsx` to use real data and have search capabilities [#4]
 - Updated `/template/[templateId]/access` page to hook it up to real data [#223]
 - Updated the Edit Question `template/[templateId]/q/[q_slug]/page.tsx` with actual data from backend and added functionality for `options` question types [#188]